Refactor filenames
